Move hooks, contexts, and more out of _core
Moving things that are definitely staying out of _core
